expand unit tests

+++ trunk/src/storage.cpp
@@ -1248,7 +1248,7 @@ ret:
 
 		// allocate a temporary, aligned, buffer
 		aligned_holder aligned_buf(aligned_size);
-		file::iovec_t b = {aligned_buf.get(), aligned_size};
+		file::iovec_t b = {aligned_buf.get(), size_t(aligned_size) };
 		size_type ret = file_handle->readv(aligned_start, &b, 1, ec);
 		if (ret < 0)
 		{
@@ -1289,7 +1289,7 @@ ret:
 
 		// allocate a temporary, aligned, buffer
 		aligned_holder aligned_buf(aligned_size);
-		file::iovec_t b = {aligned_buf.get(), aligned_size};
+		file::iovec_t b = {aligned_buf.get(), size_t(aligned_size) };
 		// we have something to read
 		if (aligned_start < actual_file_size && !ec)
 		{
@@ -1331,7 +1331,7 @@ ret:
 		, int offset
 		, int size)
 	{
-		file::iovec_t b = { (file::iovec_base_t)buf, size };
+		file::iovec_t b = { (file::iovec_base_t)buf, size_t(size) };
 		return writev(&b, slot, offset, 1);
 	}
 
@@ -1341,7 +1341,7 @@ ret:
 		, int offset
 		, int size)
 	{
-		file::iovec_t b = { (file::iovec_base_t)buf, size };
+		file::iovec_t b = { (file::iovec_base_t)buf, size_t(size) };
 		return readv(&b, slot, offset, 1);
 	}
 
@@ -2263,7 +2263,7 @@ ret:
 						m_scratch_buffer2.reset(page_aligned_allocator::malloc(m_files.piece_length()));
 
 					int piece_size = m_files.piece_size(other_piece);
-					file::iovec_t b = {m_scratch_buffer2.get(), piece_size};
+					file::iovec_t b = {m_scratch_buffer2.get(), size_t(piece_size) };
 					if (m_storage->readv(&b, piece, 0, 1) != piece_size)
 					{
 						error = m_storage->error();
@@ -2277,7 +2277,7 @@ ret:
 				// the slot where this piece belongs is
 				// free. Just move the piece there.
 				int piece_size = m_files.piece_size(piece);
-				file::iovec_t b = {m_scratch_buffer.get(), piece_size};
+				file::iovec_t b = {m_scratch_buffer.get(), size_t(piece_size) };
 				if (m_storage->writev(&b, piece, 0, 1) != piece_size)
 				{
 					error = m_storage->error();
@@ -2319,7 +2319,7 @@ ret:
 					m_scratch_buffer.reset(page_aligned_allocator::malloc(m_files.piece_length()));
 			
 				int piece_size = m_files.piece_size(other_piece);
-				file::iovec_t b = {m_scratch_buffer.get(), piece_size};
+				file::iovec_t b = {m_scratch_buffer.get(), size_t(piece_size) };
 				if (m_storage->readv(&b, piece, 0, 1) != piece_size)
 				{
 					error = m_storage->error();
